Fix rounded transfer

+++ skeleton/src/Service/Api/StripeService.php
@@ -455,8 +455,8 @@ class StripeService extends AbstractApi
                 $byMerchant[$merchant['id']] = 0;
             }
  
-            $itemTotal = (int) $item['price'] * $item['qty'] * 100;
-            $fee       = (int) $itemTotal * $this->amountFees / 100;
+            $itemTotal = (float) $item['price'] * (int) $item['qty'] * 100;
+            $fee       = $itemTotal * $this->amountFees / 100;
             $net       = $itemTotal - $fee;
  
             $byMerchant[$merchant['id']] = ($byMerchant[$merchant['id']] ?? 0) + $net;
